% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_co2.R
\name{add_co2}
\alias{add_co2}
\title{Add a the jittered range of CO2 values}
\usage{
add_co2(data, co2, jitter_amount = NULL)
}
\arguments{
\item{data}{A data frame with the column \code{companies_id}, and the list columns \code{product} and \code{company} holding the outputs at product and company level.}

\item{co2}{See \code{co2} in \code{\link[tiltIndicator:emissions_profile]{tiltIndicator::emissions_profile()}}.}

\item{jitter_amount}{Numeric. Controls the amount of noise. Passed to
\code{amount} in \code{\link[tiltIndicator:jitter_range]{tiltIndicator::jitter_range()}}.}
}
\value{
An object of the same class as \code{data}.
}
\description{
Add a the jittered range of CO2 values
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(readr)
library(tiltToyData)
library(tiltIndicatorAfter)
library(tiltAddCO2)

file <- toy_emissions_profile_products_ecoinvent()
co2 <- read_csv(file, show_col_types = FALSE)
profile <- toy_profile_emissions_impl_output()

with_co2 <- profile |>
  add_co2(co2)

with_co2 |>
  unnest_product() |>
  relocate(matches("co2"))

with_co2 |>
  unnest_company() |>
  relocate(matches("co2"))
}
