% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_noise.R
\name{summarize_noise}
\alias{summarize_noise}
\title{Summarize the mean percent noise in each gruop of jittered range of data}
\usage{
summarize_noise(data, .by = NULL)
}
\arguments{
\item{data}{A dataset with the columns min, min_jitter, max, max_jitter.}

\item{.by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}
}
\value{
See \code{\link[dplyr:summarise]{dplyr::summarize()}}.
}
\description{
Summarize the mean percent noise in each gruop of jittered range of data
}
\examples{
# styler: off
data <- tibble::tribble(
  ~min_jitter,  ~min,   ~max, ~max_jitter, ~group,
          0.8,   1.0,    2.1,         2.2,      1,
          0.9,   1.1,    2.2,         3.0,      1,

          0.1,   2.1,    3.1,         5.8,      2,
          0.2,   2.2,    3.2,         5.9,      2,
)
# styler: on

summarize_noise(data)

summarize_noise(data, .by = "group")
}
