% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltAddCO2-package.R
\docType{package}
\name{tiltAddCO2-package}
\alias{tiltAddCO2}
\alias{tiltAddCO2-package}
\title{tiltAddCO2: Add CO2}
\description{
Add CO2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/tiltAddCO2/}
  \item \url{https://github.com/2DegreesInvesting/tiltAddCO2}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltAddCO2/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
